improve import mechanic
