+++ tests/unit/ember-sinon-qunit-test.js
@@ -1,5 +1,5 @@
 import { module, skip } from 'qunit';
-import test from 'dummy/tests/ember-sinon-qunit/test';
+import test from 'ember-sinon-qunit/test-support/test';
 import assertSinonInTestContext from '../helpers/assert-sinon-in-test-context';
 
 module('Unit | ember-sinon-qunit', {
